# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Test Management Platform API
  description: |
    Comprehensive API for the Test Management Platform with Observer/Orchestrator architecture.
    
    This API provides endpoints for:
    - Authentication and user management
    - Test execution orchestration
    - Git repository management
    - Webhook integrations for CI/CD systems
    
    ## Authentication
    The API uses session-based authentication. Users must login to receive a session cookie
    that will be used for subsequent authenticated requests.
    
    ## Architecture
    The platform follows an Observer/Orchestrator pattern where the API coordinates
    external CI/CD systems rather than executing tests directly.
  version: 1.0.0
  contact:
    name: Test Management Platform
    url: http://localhost:3000
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://your-domain.com/api
    description: Production server
...
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and create session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  example: admin123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Login successful
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      description: Destroy user session
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Logout successful
...
  /auth/profile:
    get:
      tags:
        - Authentication
      summary: Get user profile
      description: Get current user profile information
      security:
        - sessionAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/change-password:
    post:
      tags:
        - Authentication
      summary: Change user password
      description: Change current user password
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - currentPassword
                - newPassword
              properties:
                currentPassword:
                  type: string
                newPassword:
                  type: string
                  minLength: 6
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Password changed successfully
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # Test Orchestration Endpoints
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session-based authentication using HTTP cookies
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: admin
        email:
          type: string
          format: email
          example: admin@example.com
        role:
          type: string
          example: admin
        created_at:
          type: string
          format: date-time
        last_login:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Invalid credentials
        code:
          type: string
          example: AUTH_FAILED
        details:
          type: object
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Authentication required
              code:
                type: string
                example: AUTH_REQUIRED
...